# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Legal Backend API",
    "description": "A comprehensive backend API for a legal services platform with client management, user administration, and authentication features.",
    "version": "1.0.0",
    "contact": {
      "name": "Legal Backend Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://your-app.vercel.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    },
    "schemas": {
...
      "ClientInput": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Client's full name",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Client's email address"
          },
          "phone": {
            "type": "string",
            "description": "Client's phone number (optional)"
          },
          "password": {
            "type": "string",
            "description": "Client's password",
            "minLength": 6
          }
        },
        "required": [
          "fullName",
          "email",
          "password"
        ]
      },
...
      "UserInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User's name",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "password": {
            "type": "string",
            "description": "User's password"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "description": "User's role",
            "default": "user"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/Client"
          }
        }
      },
...
      "ListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/Client"
                },
                {
                  "$ref": "#/components/schemas/User"
                }
              ]
            }
          }
        }
      },
...
      "BulkDeleteResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "deleted": {
                "type": "integer",
                "description": "Number of items deleted"
              }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "BlogInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Blog post title",
            "minLength": 1
          },
          "category": {
            "type": "string",
            "description": "Blog post category (optional)"
          },
          "subtitle": {
            "type": "string",
            "description": "Blog post subtitle (optional)"
          },
          "description": {
            "type": "string",
            "description": "Blog post content/description",
            "minLength": 1
          },
          "image": {
            "type": "string",
            "description": "Blog post image URL (optional)"
          },
          "published": {
            "type": "boolean",
            "description": "Publication status",
            "default": true
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
...
    }
  },
  "paths": {
    "/api/clients": {
      "post": {
        "tags": [
          "Clients"
        ],
        "summary": "Register a new client",
        "description": "Create a new client account with the provided information.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInput"
              },
              "example": {
                "fullName": "John Doe",
                "email": "john.doe@example.com",
                "phone": "+1-555-0101",
                "password": "password123"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "success": false,
                      "error": "fullName is required"
                    }
                  },
                  "invalidPassword": {
                    "value": {
                      "success": false,
                      "error": "password must be at least 6 chars"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": "email already in use"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get all clients",
        "description": "Retrieve a list of all registered clients. This is typically an admin operation.",
        "responses": {
          "200": {
            "description": "List of clients retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Delete all clients",
        "description": "⚠️ DANGER: Delete all client records. This operation cannot be undone.",
        "responses": {
          "200": {
            "description": "All clients deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                },
                "example": {
                  "success": true,
                  "data": {
                    "deleted": 42
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/blogs": {
      "post": {
        "tags": [
          "Blogs"
        ],
        "summary": "Create a blog post",
        "description": "Create a new blog post with the provided content and metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogInput"
              },
              "example": {
                "title": "Understanding Legal Rights",
                "category": "Legal Education",
                "subtitle": "A comprehensive guide for citizens",
                "description": "This blog post explains the fundamental legal rights that every citizen should know...",
                "image": "https://example.com/blog-image.jpg",
                "published": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Blog"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "success": false,
                      "error": "title is required"
                    }
                  },
                  "invalidDescription": {
                    "value": {
                      "success": false,
                      "error": "description is required"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Blogs"
        ],
        "summary": "Get all blog posts",
        "description": "Retrieve a list of all blog posts with optional filtering.",
        "parameters": [
          {
            "name": "published",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by publication status"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Limit number of results"
          },
          {
            "name": "startAfter",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Cursor for pagination - blog ID to start after"
          }
        ],
        "responses": {
          "200": {
            "description": "List of blog posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Blog"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Blogs"
        ],
        "summary": "Delete all blog posts",
        "description": "⚠️ DANGER: Delete all blog posts. This operation cannot be undone.",
        "responses": {
          "200": {
            "description": "All blog posts deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notifications": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Create a notification",
        "description": "Create a new notification with the provided content and metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationInput"
              },
              "example": {
                "title": "New Legal Update",
                "description": "Important legal information has been updated...",
                "image": "https://example.com/notification-image.jpg",
                "published": true,
                "read": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Notification"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingFields": {
                    "value": {
                      "success": false,
                      "error": "title is required"
                    }
                  },
                  "invalidDescription": {
                    "value": {
                      "success": false,
                      "error": "description is required"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Get all notifications",
        "description": "Retrieve a list of all notifications with optional filtering.",
        "parameters": [
          {
            "name": "published",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by publication status"
          },
          {
            "name": "read",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by read status"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Limit number of results"
          },
          {
            "name": "startAfter",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Cursor for pagination - notification ID to start after"
          }
        ],
        "responses": {
          "200": {
            "description": "List of notifications retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete all notifications",
        "description": "⚠️ DANGER: Delete all notification records. This operation cannot be undone.",
        "responses": {
          "200": {
            "description": "All notifications deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "description": "Create a new user account. Typically for admin operations.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              },
              "example": {
                "name": "Admin User",
                "email": "admin@example.com",
                "password": "securepass123",
                "role": "admin"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users",
        "description": "Retrieve a list of all users in the system.",
        "responses": {
          "200": {
            "description": "List of users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete all users",
        "description": "⚠️ DANGER: Delete all user records. This operation cannot be undone.",
        "responses": {
          "200": {
            "description": "All users deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/news": {
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Create a news article",
        "description": "Create a new news article with the provided content and metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsInput"
              },
              "example": {
                "title": "Legal Reform Update",
                "content": "The government has announced new legal reforms...",
                "author": "Legal Team",
                "category": "Reforms",
                "tags": [
                  "law",
                  "reform",
                  "government"
                ],
                "published": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "News article created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/News"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get all news articles",
        "description": "Retrieve a list of all news articles with optional filtering.",
        "parameters": [
          {
            "name": "published",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by publication status"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Limit number of results"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            },
            "description": "Number of results to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "List of news articles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/News"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "News"
        ],
        "summary": "Delete all news articles",
        "description": "⚠️ DANGER: Delete all news articles. This operation cannot be undone.",
        "responses": {
          "200": {
            "description": "All news articles deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/queries": {
      "post": {
        "tags": [
          "Queries"
        ],
        "summary": "Submit a legal query",
        "description": "Submit a new legal query with details and metadata.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryInput"
              },
              "example": {
                "userId": "user123",
                "title": "Contract Dispute",
                "description": "I need help with a contract dispute...",
                "category": "Contract Law",
                "priority": "high",
                "status": "open"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Query submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Query"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Queries"
        ],
        "summary": "Get all queries",
        "description": "Retrieve a list of all legal queries in the system.",
        "responses": {
          "200": {
            "description": "List of queries retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Query"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Queries"
        ],
        "summary": "Delete all queries",
        "description": "⚠️ DANGER: Delete all queries. This operation cannot be undone.",
        "responses": {
          "200": {
            "description": "All queries deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admins": {
      "post": {
        "tags": [
          "Admins"
        ],
        "summary": "Create a new admin",
        "description": "Create a new admin account with the provided information.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminInput"
              },
              "example": {
                "name": "John Doe",
                "email": "admin@example.com",
                "password": "securepass123",
                "phoneNumber": "+1234567890",
                "isActive": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "Get all admins",
        "description": "Retrieve a list of all admin accounts in the system.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            },
            "description": "Maximum number of admins to return"
          },
          {
            "name": "startAfter",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Cursor for pagination - admin ID to start after"
          }
        ],
        "responses": {
          "200": {
            "description": "List of admins retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Admin"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admins"
        ],
        "summary": "Delete all admins",
        "description": "⚠️ DANGER: Delete all admin records. This operation cannot be undone.",
        "responses": {
          "200": {
            "description": "All admins deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}